fix: Переносить Joi схему валідації юзера в окремий файл
