+++ models/users.js
@@ -1,5 +1,5 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
+// const Joi = require("joi");
 
 const userSchema = new Schema(
   {
@@ -34,19 +34,19 @@ userSchema.post("save", (error, data, next) => {
   next();
 });
 
-const addRegisterSchema = Joi.object({
-  email: Joi.string().required(),
-  password: Joi.string().required().min(6),
-});
+// const addRegisterSchema = Joi.object({
+//   email: Joi.string().required(),
+//   password: Joi.string().required().min(6),
+// });
 
-const addLoginSchema = Joi.object({
-  email: Joi.string().required(),
-  password: Joi.string().required().min(6),
-});
+// const addLoginSchema = Joi.object({
+//   email: Joi.string().required(),
+//   password: Joi.string().required().min(6),
+// });
 
-const schema = {
-  addRegisterSchema,
-  addLoginSchema,
-};
+// const schema = {
+//   addRegisterSchema,
+//   addLoginSchema,
+// };
 
-module.exports = { User, userSchema, schema };
+module.exports = { User, userSchema };
